tests, webkook response, search filter

+++ bookings/views.py
@@ -8,33 +8,29 @@ from django.conf import settings
 from django.contrib import messages
 from django.contrib.auth.decorators import login_required
 from django.contrib.auth.mixins import LoginRequiredMixin
-from django.contrib.auth.models import User
-from django.core.mail import send_mail
 from django.db.models import Q
 from django.http import HttpResponse, HttpResponseRedirect
 from django.shortcuts import get_object_or_404, redirect, render
-from django.template.loader import render_to_string
-from django.templatetags.static import static
 from django.urls import reverse, reverse_lazy
-from django.utils.html import strip_tags
 from django.utils.timezone import datetime
 from django.utils.translation import gettext_lazy as _
 from django.views.decorators.csrf import csrf_exempt
 from django.views.generic import CreateView, DetailView, UpdateView
 from django_filters import views
-from webpush import send_user_notification
 
 from apartments.models import Apartment
 from bookings.utils import booking_dates_assignment, get_next_prev_month
-from home.models import AdminEmail, PhoneSnippet
+from home.models import PhoneSnippet
 
 from .filters import BookingsFilter
 from .forms import (
-    BookingForm, BookingUpdateForm, OnlineBookingDetailsForm,
+    BookingForm,
+    BookingUpdateForm,
+    OnlineBookingDetailsForm,
     OnlineBookingForm,
 )
 from .models import Booking
-from .utils import calculated_price, handle_error_notification
+from .utils import calculated_price, handle_error_notification, WebhookResponse
 
 logger = logging.getLogger("django")
 stripe.api_key = settings.STRIPE_SECRET_KEY
@@ -63,7 +59,6 @@ def booking_search(request, year=None, month=None):
     context = {"form": OnlineBookingForm()}
 
     if "submit" in request.GET:
-        # TODO REMOVE
         form = OnlineBookingForm(request.GET)
         if form.is_valid():
             arrival = datetime.strptime(
@@ -82,7 +77,10 @@ def booking_search(request, year=None, month=None):
                 if session_email:
                     qs = Booking.objects.bookings_periods(
                         apartment, arrival, departure
-                    ).exclude(email=session_email)
+                    ).exclude(
+                        Q(email=session_email)
+                        & Q(stripe_transaction_status="pending")
+                    )
                 else:
                     qs = Booking.objects.bookings_periods(
                         apartment, arrival, departure
@@ -126,17 +124,15 @@ def calendars(request, year=None, month=None):
 
     cal_months = get_next_prev_month(year, month)
 
-    ap1 = Apartment.objects.get(name="1")
-    ap2 = Apartment.objects.get(name="2")
-    ap3 = Apartment.objects.get(name="3")
-    ap4 = Apartment.objects.get(name="4")
-
-    ap1_bookings_dict = booking_dates_assignment(ap1, year, month)
-    ap2_bookings_dict = booking_dates_assignment(ap2, year, month)
-    ap3_bookings_dict = booking_dates_assignment(ap3, year, month)
-    ap4_bookings_dict = booking_dates_assignment(ap4, year, month)
+    all_apartments = Apartment.objects.all().order_by("name")
+    apartments_booking_dates = {}
+    for num, apartment in enumerate(all_apartments):
+        apartments_booking_dates[
+            "ap{0}_dates".format(num + 1)
+        ] = booking_dates_assignment(apartment, year, month)
 
     context = {
+        **apartments_booking_dates,
         "year": year,
         "month": month,
         "displayed_month": date(year, month, 1),
@@ -146,10 +142,6 @@ def calendars(request, year=None, month=None):
         "next_year": cal_months["next_year"],
         "num_days": range(1, num_days + 1),
         "first_day": first_day,
-        "ap1_dates": ap1_bookings_dict,
-        "ap2_dates": ap2_bookings_dict,
-        "ap3_dates": ap3_bookings_dict,
-        "ap4_dates": ap4_bookings_dict,
     }
 
     return render(
@@ -242,8 +234,10 @@ class BookingUpdateView(LoginRequiredMixin, UpdateView):
 @login_required(login_url="users_app:user-login")
 def delete_booking(request, pk):
     """Delete booking."""
-    booking_to_delete = get_object_or_404(Booking, id=pk)
-    booking_to_delete.delete()
+
+    if request.method == "POST":
+        booking_to_delete = get_object_or_404(Booking, id=pk)
+        booking_to_delete.delete()
     return HttpResponseRedirect(reverse("bookings_app:bookings"))
 
 
@@ -382,7 +376,7 @@ def onlinebooking(request, arrival=None, departure=None, pk=None):
 
 def success(request):
     for key in list(request.session.keys()):
-        if not key.startswith("_"):  # skip keys set by the django system
+        if key == "email":  # skip keys set by the django system
             del request.session[key]
     # stripe.api_key = settings.STRIPE_SECRET_KEY
     # checkout_session_id = request.GET.get('session_id', None)
@@ -415,111 +409,58 @@ def stripe_webhook(request):
         logger.error(f"Stripe-webhook error: {repr(e)}")
         return HttpResponse(status=400)
     except Exception as e:
-        error_subj = "Webhook error"
-        error_msg = f"{repr(e)}"
-        handle_error_notification(error_subj, error_msg)
+        err_subj = "Webhook error"
+        err_msg = f"{repr(e)}"
+        handle_error_notification(err_subj, err_msg)
         return HttpResponse(status=400)
 
-    if event["type"] == "checkout.session.completed":
+    try:
         session = event["data"]["object"]
         session_id = session.get("id", None)
+    except KeyError as e:
+        err_subj = "Webhook sent, but event['data']['object'] not found"
+        err_msg = err_subj + "in it's content."
+        handle_error_notification(err_subj, err_msg)
 
+    if event["type"] == "checkout.session.completed":
         try:
             booking = Booking.objects.get(stripe_checkout_id=session_id)
         except Booking.DoesNotExist:
             booking = None
-            error_subj = str(_("Session completed, booking not found"))
-            error_msg = f"Session completed (session id: {session_id}) \
-                webhook was sent, but related booking was not found \
-                in the database."
-            handle_error_notification(error_subj, error_msg)
+            err_subj = str(_("Session completed, booking not found"))
+            err_msg = (
+                f"Session completed (session id: {session_id}) "
+                "webhook was sent, but related booking was not found "
+                "in the database."
+            )
+            handle_error_notification(err_subj, err_msg)
 
         if booking and session.payment_status == "paid":
             booking.stripe_transaction_status = "success"
             booking.paid = True
             booking.save()
 
-            from_email = settings.EMAIL_HOST_USER
-            hotel_email = AdminEmail.objects.last()
-            subject = f"Rezerwacja Ap. nr {booking.apartment.name} \
-                od {booking.date_from}"
-            notes = booking.notes if booking.notes else "-"
-            url = str(
-                request.build_absolute_uri(
-                    reverse("bookings_app:booking", kwargs={"pk": booking.id})
-                )
+            return WebhookResponse(
+                request=request, booking=booking, status=200
             )
-            msg = f"""Nowa rezerwacja: #{booking.id} \n
-    Apartament nr {booking.apartment.name} \n
-    od {booking.date_from} do {booking.date_to} \n
-    Gość: {booking.guest} \n
-    Uwagi gościa: {notes} \n
-    utworzona: {booking.created_at} \n
-    Zobacz: {url}"""
-
-            # send email to hotel
-            try:
-                send_mail(subject, msg, from_email, [hotel_email])
-            except Exception as e:
-                error_subj = "Confirmation email not sent for booking"
-                error_msg = f"Confirmation email not sent for booking num \
-                    {booking.id}. {repr(e)}"
-                handle_error_notification(error_subj, error_msg)
-
-            # send webpush notification to hotel
-            try:
-                hotel = User.objects.filter(email=hotel_email).last()
-                payload = {
-                    "head": subject,
-                    "body": f"Nowa rezerwacja: #{booking.id}",
-                    "url": url,
-                    "icon": static("img/favicon/android-chrome-96x96.png"),
-                }
-                send_user_notification(user=hotel, payload=payload, ttl=1000)
-            except User.DoesNotExist:
-                hotel = None
-                logger.error("Notification not sent. User doesn't exist.")
-            except Exception as e:
-                handle_error_notification(
-                    "Error while sending webpush confirmation to hotel",
-                    repr(e),
-                )
-
-            # send email to guest
-            try:
-                hotel_phone = PhoneSnippet.objects.last()
-                subject = "Potwierdzenie rezerwacji B4B"
-                html_message = render_to_string(
-                    "bookings/confirmation-email.html",
-                    {"booking": booking, "phone": hotel_phone},
-                )
-                plain_message = strip_tags(html_message)
-                to = booking.email
-                send_mail(
-                    subject,
-                    plain_message,
-                    from_email,
-                    [to],
-                    html_message=html_message,
-                )
-            except Exception as e:
-                error_subj = "Sending confirmation email failed"
-                error_msg = (
-                    f"Confirmation email not sent for {booking}. {repr(e)}"
-                )
-                handle_error_notification(error_subj, error_msg)
 
     elif event["type"] == "checkout.session.expired":
-        session = event["data"]["object"]
-        session_id = session.get("id", None)
         try:
             booking = Booking.objects.get(stripe_checkout_id=session_id)
             booking.delete()
         except Booking.DoesNotExist:
             logger.error(
-                f"""Session expired (session id: {session_id})
-                webhook was sent,
-                but related booking was not found in the database.
-                """
+                f"Session expired (session id: {session_id}) "
+                "webhook was sent, but related booking was "
+                "not found in the database."
             )
+
+    else:
+        error_subj = "Webhook with unexpected event type received."
+        error_msg = (
+            "Wehook was sent, but contains unexpected event type: "
+            f"{event['type']}. Session id: {session_id}"
+        )
+        handle_error_notification(error_subj, error_msg)
+
     return HttpResponse(status=200)
